requirements/reqs.txt
# pip requirements for the stuff buzzfire runs on

redis
oauth2
django-extensions